replace old style cast by static_cast

+++ rgbd/tools/record_to_video.cpp
@@ -86,7 +86,8 @@ int main(int argc, char **argv) {
             if (!initialized)
             {
                 // If not, do so
-                video_size = cv::Size2i(size * rgb_image.cols, size * rgb_image.rows);
+                video_size = cv::Size2i(static_cast<int>(size * rgb_image.cols),
+                                        static_cast<int>(size * rgb_image.rows));
                 video_writer.open(filename.c_str(), cv::VideoWriter::fourcc(format[0], format[1], format[2], format[3]), video_frame_rate, video_size);
 
                 if (!video_writer.isOpened())
